nlp/project/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dataclass-wizard==0.35.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-core==1.0.1
langchain-openai==1.0.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
tiktoken==0.12.0
todoist_api_python==3.1.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wikipedia==1.4.0
xxhash==3.6.0
zstandard==0.25.0
